try SunkenSource
this gets around a pretty specific usecase where we  don't know when the
messages will arive on a normal Source() and we would have to have a lot
of blocking src.Next() calls that we would have to manage somehow.

The passed source gets a raw codec.Packet and should check the flags
before reading the body.