try ByteSource (updates #8)
This tries the first suggesting of an sql.Rows like source.

I feel like to reduce the internal buffering even more we have to split
up muxrpc/codec, though. Right now it reads the header and body in one take.

If we want to reduce the buffering there we can only read the header and
then need a ReadInto([]byte) on the source of the source directly.

This however needs the handling source to be ready immediatly when
packets arrive and this doesn't sound reasonable, especially for the
verification pipeline in ssb.

I will follow up with some benchmarks for memory consumtion but this seems
much nicer than an dynamic number of codec.Packt (with the []byte Body)
objects in luigi.Pipes.