add sleep and document problem around NoSuchMethod error
It's expected behavior to get an error from the methods on Endpoint (so
Async, Source, Sink, Duplex) if the remote doesn't known/server a method
by that name.

However, since the protocol doesn't have an acknowledgment when opening
streams there either is an error packet if the method is unknown or no
packet at all until the first data frame is received.

This leaves room for uncertainty, which is always bad in programming. I
kludged this problem with a sleep which can be a problem on long network
delays. On the plus side, the EndErr packet is send rather quickly.

Also, this _just_ affects code depending on the error from the call. If
this would glitch (the remote takes longer to send EndErr then the
sleep) the stream would yield no elements and return the err from
src.Err().

(this affects feature negotiation in go-ssb between legacy and ebt
replication)